+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.5.4] - 2025-10-02
+### Fixed
+- Add status validation to prevent unnecessary metadata updates
+- Fixed location id in melidata tracks for admin pages
+- Fix order-pay form validation error with Checkout Transparent payment method
+### Added
+- Improve plugin checkout metrics
+
 ## [8.5.3] - 2025-09-29
 ### Fixed
 - Wait for security code to be updated before submitting the form on fast payments flow